Reduce the number of data key variations
Recorders that include a time element in the data key can produce a
large number of variations. Thse recorders already reduced the number
of unique keys using "record unit" calculations, but the granularity of
data necessary depends on the current duration. For smaller durations
(ten seconds or under) a 1 or 0.5 second duration may be preferred.
Using such a small value for auras that last 30 minutes would create
lots of noise in the data.